GERT 2
V2 of GERTRUDE: Stable version and managed to get pnumber a fixed number on the project database instead of it being a number for episodes/episode folders, saving space in the database. Also introduced small changes such as making pnumber an option in editproject and making a color used the done/undone/blame embeds for a specific project (I forgor the readme and doc are different if uncapitalized).