Gert v2.0.1
Editproject now is able to edit everything from Newproject, so if you mess up, you are now be able edit it to the correct value. Also, fixed bug for Editproject in where if a wrong string/number was inputed, it would be put into the database anyways instead of being rejected.